Refactor ABI fixup generation logic
Currently, it is hardcoded to support only a very specific scenario where an instance method that returns a value by reference is being detoured to a static method.

However, with this rewrite in mind, it should be much easier to introduce new conditions for generating ABI fixups in the future.

Additionally, I have added [quite] a bit more information on why and when these fixups are needed in the first place.

This change does **not** alter the existing behavior in **any** way.